backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
feedparser
redis
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pymupdf
google-generativeai
groq
# sentence-transformers  # uncomment when needed for semantic matching (Task 8)
pgvector
websockets
pydantic
pydantic-settings
httpx
python-dotenv
bcrypt==4.0.1
